[IDE] Avoid macro expansions in `getEquivalentDeclContextFromSourceFile`
Unfortunately haven't been able to come up with a test case for this,
but there seem to be cases where we're incorrectly picking up
a macro-expanded accessor from the cached AST when searching for the
original decl. Make sure we only consider decls that have been
written by the user.

rdar://151926231