[OSSACanOwned] Dead-end destroy not end liveness.
The utility operates on the copy-extended lifetime of a value.  In
non-dead-end region, the last destroy seen on any path must end that
copy-extended lifetime.  In dead-end regions, however, while we continue
to lack complete lifetimes, the final destroy may not actually end the
copy-extended lifetime: copies may leak into the dead-end past that
destroy.

When extending lifetimes into dead-ends, the utility constructs a
second, temporary instance of liveness to which the final destroys are
added.  This temporary instance is used to discover which dead-ends
the lifetime extends into.  The second, temporary, destroy-extended
instance of liveness is necessary because it includes the full
copy-extended lifetime of the value--away from dead-ends--which enables
discovering dead-ends which are part of the copy-extended lifetime.

Previously, all destroys which were the final destroys on some path were
added to that second, temporary instance of liveness.  But the means by
which it is determined whether a destroy is a "final destroy" is just by
checking whether it is within the boundary of the second, temporary
liveness instance.  This is correct for destroys in non-dead-end
regions.  But because copies may leak into the dead-end past, a destroy
seen in a dead-end region, the last destroy we happen to see in a
dead-end region may not be the final destroy.

Here, destroys in dead-end regions are not added to the second,
temporary liveness instance.  The result is to include the full dead-end
region in which such destroys appear as part of the copy-extended
lifetime of the value.

rdar://154476285