6.2: [Test] Re-disable test in asserts builds.
The test was reenabled for the `swift_stdlib_asserts` context in
https://github.com/swiftlang/swift/pull/82559/ .  That had the effect of
enabling it for noasserts builds.  But the test uses subpass flags which
are only heeded in asserts builds.  That behavior is disabled--the flags
are made always effective--in
https://github.com/swiftlang/swift/pull/82736 .  But there's no reason
to take that change onto a release branch.  Just disable the test again
in exactly the relevant configuration.

rdar://154499140