[SwiftCompilerSources] Explicitly construct `StringRef` from a literal when passing it to `==`
There are multiple types that get imported from C++ as conforming
to `ExpressibleByStringLiteral`: `StringRef` and `StaticString` among them.

This results in ambiguity when using `==` operator because overloads
are structured in a way that accepts both `StringRef` and `StaticString`
on both sides, so for cases like:

```swift
func test(lhs: StringRef) {
  lhs == "<<test>>"
}
```

There are multiple solutions now because "<<test>>" could be passed
to both `StringRef` and `StaticString` overload.

The type-checker used to pick `==(StringRef, StringRef)` overload in
this case because it has homogenous parameter types but this is no
longer the case because this behavior was too aggressive and led to
sub-optimal choices by completely skipping other viable overloads.

Resolves: rdar://154719565